Refactor: add MusicLoginButton, MusicSearchForm, MusicTrackItem, Remove: logo.png
